feat: fix menu
